+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed #163: Was detecting std lib packages when the GOROOT was different at
   runtime than compile time.
+- Fixed #165: glide update panics with --no-recursive option.
 
 # Release 0.8.0 (2015-12-10)